Delete .gitattributes
A description of what I did was that I had created the data table of the planets. I was a little confused because it didn't turn out exactly like it did on the video and followed the instructions but, I got it almost to the same as it was compared to the video.